refactor: update logger imports
