+++ services/data/chart-data-service.ts
@@ -12,7 +12,7 @@ import { IChartDataService } from './interfaces';
 import { OHLCData, Timeframe } from '@/types/chart';
 import { ExchangeType, ProductType } from '@/types/api';
 import { useRootStore } from '../../store/rootStore';
-import { logger } from '../../utils/logger';
+import { logger } from '@/utils/common';
 import { normalizeSymbol } from '../../utils/formatters';
 import { getSocketService } from '../socket/index';
 import { cacheService } from '../cache/service';
