+++ services/socket/bitget-integration.ts
@@ -8,7 +8,7 @@
 
 import { ProductType } from '@/types/api';
 import { BitgetApiClient } from '../api/bitget/client.new';
-import { logger } from '../../utils/logger';
+import { logger } from '@/utils/common';
 import { IBitgetIntegration } from './interfaces';
 
 /**
