+++ services/socket/mock-service.ts
@@ -10,7 +10,7 @@ import { EventEmitter } from 'events';
 import { OrderBookData, OrderBookEntry } from '@/types/common/orderbook';
 import { OHLCData, Timeframe } from '@/types/chart';
 import { ProductType } from '@/types/api';
-import { logger } from '../../utils/logger';
+import { logger } from '@/utils/common';
 import { ISocketService } from './interfaces';
 
 /**
