+++ services/socket/service.ts
@@ -8,7 +8,7 @@ import { Socket } from 'socket.io-client';
 import { OrderBookData } from '@/types/market';
 import { OHLCData, Timeframe } from '@/types/chart';
 import { ProductType } from '@/types/api';
-import { logger } from '../../utils/logger';
+import { logger } from '@/utils/common';
 import { ISocketService } from './interfaces';
 
 /**
